💄 change globals colors

+++ events/guild_messages/messageCreate.js
@@ -1,4 +1,4 @@
-const { MessageEmbed } = require('discord.js')
+const { MessageEmbed, Permissions } = require('discord.js')
 const langFr = require('../../languages/fr/events/messageCreate.json')
 const langEn = require('../../languages/en/events/messageCreate.json')
 
@@ -14,13 +14,15 @@ module.exports = {
 
     const guildBadWords = guild.badWords
 
-    for (const word of guildBadWords) {
-      if (message.content.includes(word)) {
-        message.channel.send(`<@${message.author.id}>\n${lang.autoModErreur1} \`${word}\` ${lang.autoModErreur2}`)
-        setTimeout(() => {
-          message.delete()
-        }, 5000)
-        return
+    if (!message.member.permissions.has([Permissions.FLAGS.ADMINISTRATOR])) {
+      for (const word of guildBadWords) {
+        if (message.content.includes(word)) {
+          message.channel.send(`<@${message.author.id}>\n${lang.autoModErreur1} \`${word}\` ${lang.autoModErreur2}`)
+          setTimeout(() => {
+            message.delete()
+          }, 5000)
+          return
+        }
       }
     }
